refactor News And Comment Screen Ui
