+++ src/screen/drawerScreen/channels/index.tsx
@@ -5,6 +5,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import Channel from './Channel';
 import ChannelDetail from './ChannelDetail';
 import NewsDetail from '../../../component/newsDetail';
+import Comment from '../../../component/comment';
 
 import { DrawerProps } from '../../../interfaces';
 
@@ -12,6 +13,7 @@ type RootStackParamList = {
   Channel: undefined;
   ChannelDetail: undefined;
   NewsDetail: undefined;
+  Comment: undefined;
 };
 
 const Stack = createStackNavigator<RootStackParamList>();
@@ -28,6 +30,7 @@ export default function Index(props: DrawerProps) {
         <Stack.Screen name='Channel' component={Channel} />
         <Stack.Screen name='ChannelDetail' component={ChannelDetail} />
         <Stack.Screen name='NewsDetail' component={NewsDetail} />
+        <Stack.Screen name='Comment' component={Comment} />
       </Stack.Navigator>
     </SafeAreaView>
   );
